move newVideo to video.go
